Remove prim function keywords
This takes the parser out of the main path for
most builtin functions, and frees up the keywords
to be normal values. Which means things like `sin`
can now be shadowed correctly.

To do this, we need to wire in the primitive functions
differently. So we wire in a normal dictionary function
for each prim, and load it when opening the prelude.