Fix substitutions in queries when checking functions.
By only using the top level substitutions in the funciton
arguments, we could mistakenly delete a temporality
or possibility argument in the query itself when running
the final ad-hoc transformation

```
forall t : Temporality. t a ~ a
```

when really, `t` should have been substituted for Aggregate
or Element.

This could break the type system, and allow for queries
requiring two passes over the data to type check.