Bump up Zebra and Viking.
The story is that ResourceT 1.1.* doesn't support exceptions > 0.8. It doesn't
implement a required method (it still builds though...).

The downside is that ResourceT 1.2.* is broken, and doesn't have MonadBaseControl
instances, which we were using through lifted-async.

So we've added a newtype wrapper for ResourceT with their old defintions of
MonadBase / MonadBaseControl.