Add sorbet test suites to the cabal file.
I've rewired the time projection primitives to be like the standard
functions, which are parsed as variables; and re-added introduced
functions back to the source generator.

This required a small transform to be written in the Parse Render
roundtrip tests, as prims are parsed to a variable with the same
pretty-print.