Change data flow in dead code pass to halve total compilation time.
Previously, about 60% of total compilation time in test dictionaries
was spent doing set unions, as if statements would be analysed with
the same set then have their results unioned.

This was correct, but, as the statements can be numerous and tiny,
doing this when there was a huge number of variables just leads to
a lot of work done to rejoin the results.

But we don't need to, we can run the blocks as if they were sequential
without any issue, as both are scoped and don't shadow, so locally
bound variables will be used and unscoped, or removed in place, and
accumulators will be handled correctl too.

This speeds up this pass by orders of magnitude.