Improve annotations for applications in source language.
Previously, the annotations with App constructors were a bit of
a lie, as the left hand side was always the same as the outer
application.

This meant that our types shown in the language server when
hovering were incorrect, and was just a bit unfortunate, as the
true types for functions were lost.

Fortunately, this is mostly a simplification, as App nodes no
longer need to do lookups for type schemes, and can focus on
just themselves. To maintain our "no closures" invariant, we
can lift our anyArrows check to a simple, non-traversing one.

Some work was required on the Reification phase and ToCore phase,
as they made assumed the previous behaviour.

Also remove a bunch of redundant checks in convertReduce, which
were looking for pure expressions to bind as precomputations; all
of these were already handled in the very first pattern.