Reduce memory usage by recognsising and deleting unnecessary copies.
We elaborate immutable Core functions to avalanche and insert extra
Copy operations when updating Arrays (which are usually map ops).

This maintains correct semantics when we compile to C and have Arrays
as allocated values on the heap with pointers to them.

Unfortunately though, some of these copy operations were turning up
in our loops, which doesn't make a lot of sense, especially for
queries like

```
from ... in group ... in ...
```

We fix this by removing copy operations which can't affect the results
by tracking aliases which might point the arrays, as well as calculating
variable usage for future statements.

When we encounter a copy operation, if no future statements use a variable
which could point the same memory location as the value about to be copied,
then we can omit the copy step completely.