Improve Let Scan type checking.
Due to how core is elaborated, we can only scan aggregations which
are themselves produced in the context; that is, you can't use a
predefined aggregate.

Also, the initial value in the let scan, like the initial value in
an element let binding, can't actually be used, so we can reduce
the amount of code produced by just embedding an error a the initial
state.