Fix core elaboration errors in nested folds with Latest.
These are some pretty long standing issues, but were very unlikely
to be encountered when the only way to bind Element values was with
let.

Now that we have Filter Let and Scan contexts, it's more likely, and
we need to make sure that we don't miscompile using these contexts.

Previously, Element values weren't being repacked in Folds which
used Latest. Top level Latest contexts were fine, as they do indeed
repack inside the "input" environment.

This is a different, and probably cleaner approach, where the
conversion state keeps track of binding which are Element temporality,
and knows that it should pack and rebind them when we hit a Latest
context.

This also fixes a long standing bug in destructPairs, whcih was
getting the associativity wrong when the number of arguments was
large.