Fix for graph search and usage set.
These two bugs serendipitously cancelled each other out in most
use cases, but it's good to fix them.

The Graph Search should include the root terms (they could be used).

When we Write to an accumulator, we shouldn't retain usages of it
from after that right, as the version written is different to the
version before hand, so a usage of the latter value is not a usage
of the current one.