Permit functions to demand pure Temporal modes.
Because you can write a function which takes the initial value
passed to a fold, we should allow functions to require a pure
temporality. Previously we had an extra _ignore_ mode on the
join if both expected and provided types, but now we lub these
and require they lift to the expected mode.

Also improve type inference correctness for possibilities when
definite modes are demanded. We don't enforce that a definite
value is provided, but instead lift up the possible mode into
the result if it is. This makes type inference more correct
and consistent with the type inferred after inlining.