Asymmetric unification for higher order modal types.
Because we can allow unadorned functions to written (using subsumption
and type annotations) we need to be a bit careful about what we permit
in terms of accepted higher order arguments. So we adopt an asymmetric
approach, where pure (unadorned) modes accepted is fine, but required
forces a Pure equality.

There's still one or two holes left.